+++ packages/core/src/components/plate/Plate.spec.tsx
@@ -93,4 +93,77 @@ describe('Plate', () => {
       ]);
     });
   });
+
+  describe('when renderAboveSlate renders null', () => {
+    it('should not normalize editor children', () => {
+      const plugins: PlatePlugin[] = [
+        {
+          key: 'a',
+          renderAboveSlate: () => {
+            return null;
+          },
+        },
+      ];
+
+      const editor = createPlateEditor({
+        plugins,
+      });
+
+      expect(() =>
+        render(<Plate editor={editor} initialValue={[{}] as any} />)
+      ).not.toThrowError();
+    });
+  });
+
+  describe('when renderAboveSlate renders children', () => {
+    it("should not trigger plugin's normalize", () => {
+      const plugins: PlatePlugin[] = [
+        {
+          key: 'a',
+          renderAboveSlate: ({ children }) => {
+            return <>{children}</>;
+          },
+        },
+      ];
+
+      const editor = createPlateEditor({
+        plugins,
+      });
+
+      expect(() =>
+        render(<Plate editor={editor} initialValue={[{}] as any} />)
+      ).toThrowError();
+    });
+  });
+
+  describe('when nested Plate', () => {
+    it('should work', () => {
+      const plugins: PlatePlugin[] = [
+        {
+          key: 'a',
+          isElement: true,
+          isVoid: true,
+          component: ({ children, attributes }) => (
+            <div {...attributes}>
+              <Plate id="test" />
+              {children}
+            </div>
+          ),
+        },
+      ];
+
+      const editor = createPlateEditor({
+        plugins,
+      });
+
+      expect(() =>
+        render(
+          <Plate
+            editor={editor}
+            initialValue={[{ type: 'a', children: [{ text: '' }] }] as any}
+          />
+        )
+      ).not.toThrowError();
+    });
+  });
 });