+++ packages/core/src/types/plate/PlateEditor.ts
@@ -1,7 +1,8 @@
 import { KeyboardEvent } from 'react';
+import { Path } from 'slate';
 import { TEditor, Value } from '../../slate/editor/TEditor';
 import { THistoryEditor } from '../../slate/history-editor/THistoryEditor';
-import { TRange } from '../../slate/index';
+import { EElement, TElement, TRange } from '../../slate/index';
 import { TReactEditor } from '../../slate/react-editor/TReactEditor';
 import { WithPlatePlugin } from '../plugin/PlatePlugin';
 import { PluginKey } from '../plugin/PlatePluginKey';
@@ -14,5 +15,18 @@ export type PlateEditor<V extends Value = Value> = TEditor<V> &
     plugins: WithPlatePlugin<{}, V>[];
     pluginsByKey: Record<PluginKey, WithPlatePlugin<{}, V>>;
     prevSelection: TRange | null;
+
+    /**
+     * Default block factory.
+     * @default [{ type: getPluginType(editor, ELEMENT_DEFAULT), children: [{ text: '' }] }]
+     */
+    blockFactory: (node?: Partial<TElement>, path?: Path) => EElement<V>;
+
+    /**
+     * Editor children factory.
+     * @default [editor.blockFactory()]
+     */
+    childrenFactory: () => V;
+
     currentKeyboardEvent: KeyboardEvent | null;
   };