+++ packages/nodes/link/src/transforms/submitFloatingLink.ts
@@ -20,10 +20,13 @@ import { upsertLink } from './index';
 export const submitFloatingLink = <V extends Value>(editor: PlateEditor<V>) => {
   if (!editor.selection) return;
 
-  const { isUrl } = getPluginOptions<LinkPlugin, V>(editor, ELEMENT_LINK);
+  const { isUrl, forceSubmit } = getPluginOptions<LinkPlugin, V>(
+    editor,
+    ELEMENT_LINK
+  );
 
   const url = floatingLinkSelectors.url();
-  const isValid = isUrl?.(url);
+  const isValid = isUrl?.(url) || forceSubmit;
   if (!isValid) return;
 
   const text = floatingLinkSelectors.text();
@@ -35,6 +38,7 @@ export const submitFloatingLink = <V extends Value>(editor: PlateEditor<V>) => {
     url,
     text,
     target,
+    isUrl: (_url) => (forceSubmit || !isUrl ? true : isUrl(_url)),
   });
 
   setTimeout(() => {