+++ packages/selection/src/index.ts
@@ -4,5 +4,8 @@
 
 export * from './blockSelectionStore';
 export * from './createBlockSelectionPlugin';
+export * from './onChangeBlockSelection';
+export * from './useHooksBlockSelection';
 export * from './components/index';
+export * from './queries/index';
 export * from './utils/index';